updated fields: del, act, upd

+++ projects/ng-common/src/lib/services/backend-http-service-base.ts
@@ -1,14 +1,17 @@
 import { HttpClient } from '@angular/common/http';
-import { ObjectWithId, BackendServiceBase, ApiListResult, ApiResult, ApiBatchProcess, ApiBatchResult, DbQuery, ObjectHelper, ApiStatus, ConnectTo } from 'ts-common'
-import { instanceToPlain, plainToInstance } from "class-transformer";
-import { Observable, map, Subject, take } from "rxjs";
+import { ObjectWithId, BackendServiceBase, ApiListResult, ApiResult, ApiBatchProcess, ApiBatchResult, DbQuery, ObjectHelper, ApiStatus, ConnectTo, QueryCondition, QueryOperator, ArrayHelper } from 'ts-common'
+import { Type, instanceToPlain, plainToInstance } from "class-transformer";
+import { Observable, map, Subject, take, of } from "rxjs";
 import { Firestore, collection, collectionData, addDoc, CollectionReference, updateDoc, serverTimestamp, doc, setDoc } from '@angular/fire/firestore';
 import { inject } from '@angular/core';
+import * as _ from "lodash";
+import { TypeInfo } from 'ts-altea-model';
 
 export enum ObjectChangeType {
   create = 'create',
   update = 'update',
-  delete = 'delete'
+  delete = 'delete',
+  softDelete = 'softDelete'
 }
 
 export class ObjectChange<T extends ObjectWithId> {
@@ -30,10 +33,48 @@ export class ObjectChange<T extends ObjectWithId> {
   }
 }
 
+export class ObjectCache<T extends ObjectWithId> {
+  objects: T[] = []
+
+  @Type(() => Date)
+  lastSync: Date
+
+
+  updateObjectsInCache(objects: T[]): boolean {
+
+    if (!ArrayHelper.AtLeastOneItem(objects))
+      return false
+
+    let atLeastOneChange = false
+
+    for (let object of objects) {
+
+      if (!object.id)
+        continue
+
+      let idx = this.objects.findIndex(obj => obj.id == object.id)
+
+      if (idx >= 0)
+        this.objects[idx] = object
+      else
+        this.objects.push(object)
+
+      atLeastOneChange = true
+    }
+
+    return atLeastOneChange
+
+  }
+
+
+
+}
 
 
 export class BackendHttpServiceBase<T extends ObjectWithId> extends BackendServiceBase<T> {   //  extends BaseObject
   protected firestore: Firestore = inject(Firestore)
+
+  protected softDelete = false
 
   changes$: Subject<ObjectChange<T>> = new Subject<ObjectChange<T>>()
 
@@ -45,9 +86,14 @@ export class BackendHttpServiceBase<T extends ObjectWithId> extends BackendServi
    * @param http 
    * @param firestoreDocUrl the full path of a firestore document that will be updated whenever this object type is updated
    */
-  constructor(protected type: { new(): T; }, private host: string, private urlDifferentiator: string, private http: HttpClient, private firestoreDocUrl?: string) {
+  constructor(protected type: { new(): T; }, protected typeName: string, private host: string, private urlDifferentiator: string, private http: HttpClient, private firestoreDocUrl?: string) {
     super()
+
+    this.changes$.subscribe(objectChange => this.updateCache(objectChange))
   }
+
+
+
 
   async post$<T>(url: string, body: any): Promise<T> {
     const me = this
@@ -72,6 +118,19 @@ export class BackendHttpServiceBase<T extends ObjectWithId> extends BackendServi
   }
 
   get(id: string, includes: string | null = null): Observable<T> {
+
+    if (this.caching) {
+
+      
+      let cachedObject = this.cache.objects.find(obj => obj.id == id)
+
+      if (cachedObject) {
+        console.warn('From cache:', cachedObject)
+        return of(cachedObject)
+      }
+        
+
+    }
 
     const queryString = []
 
@@ -101,10 +160,7 @@ export class BackendHttpServiceBase<T extends ObjectWithId> extends BackendServi
       me.get(id, includes).pipe(take(1)).subscribe(res => {
         resolve(res)
       })
-
-
     })
-
   }
 
 
@@ -163,7 +219,8 @@ export class BackendHttpServiceBase<T extends ObjectWithId> extends BackendServi
       if (res && res.status === ApiStatus.ok) {
         console.warn('triggering changes$ !')
 
-        this.changes$.next(new ObjectChange(object, isSoftDelete ? ObjectChangeType.delete : ObjectChangeType.update))
+        // notify about change & update local cache if needed: 
+        this.changes$.next(new ObjectChange(object, isSoftDelete ? ObjectChangeType.softDelete : ObjectChangeType.update))
 
         this.updateFirestore()
       }
@@ -243,22 +300,39 @@ export class BackendHttpServiceBase<T extends ObjectWithId> extends BackendServi
 
 
   delete(objectId: string): Observable<ApiResult<any>> {
-    const res = this.http.delete<any>(`${this.host}/${this.urlDifferentiator}/${objectId}`).pipe(map(res => {
 
-      res = this.makeApiResultTyped(res)
+    if (this.softDelete) {
 
-      if (res.status == ApiStatus.ok) {
-        this.changes$.next(new ObjectChange<T>(objectId, ObjectChangeType.delete))
-        this.updateFirestore()
+      let softDelete = {
+        id: objectId,
+        deleted: true,
+        deletedAt: new Date()
       }
+
+      return this.update(softDelete, true)
+
+
+    } else {
+
+      const res = this.http.delete<any>(`${this.host}/${this.urlDifferentiator}/${objectId}`).pipe(map(res => {
+
+        res = this.makeApiResultTyped(res)
+
+        if (res.status == ApiStatus.ok) {
+
+          // notify about change & update local cache if needed: 
+          this.changes$.next(new ObjectChange<T>(objectId, ObjectChangeType.delete))
+          this.updateFirestore()
+        }
+
+        return res
+      }
+      ))
+
+      //this.changes$.next(new ObjectChange<T>(objectId, ObjectChangeType.delete))
 
       return res
     }
-    ))
-
-    //this.changes$.next(new ObjectChange<T>(objectId, ObjectChangeType.delete))
-
-    return res
   }
 
   delete$(objectId: string): Promise<ApiResult<T>> {
@@ -314,7 +388,14 @@ export class BackendHttpServiceBase<T extends ObjectWithId> extends BackendServi
   }
 
 
-  query(query: DbQuery): Observable<ApiListResult<T>> {
+  query(query: DbQuery, useCache: boolean = true): Observable<ApiListResult<T>> {
+
+    if (useCache && this.caching) {
+      console.warn('Getting data from cache!')
+      let objects = this.queryFromCache(query)
+      return of(new ApiListResult<T>(objects))
+    }
+
 
     console.warn(query)
 
@@ -333,13 +414,23 @@ export class BackendHttpServiceBase<T extends ObjectWithId> extends BackendServi
     ))
   }
 
-  query$(query: DbQuery): Promise<T[]> {
+  query$(query: DbQuery, useCache: boolean = true): Promise<T[]> {
+
+
+
 
     const me = this
 
     return new Promise<any>(function (resolve, reject) {
 
-      me.query(query).pipe(take(1)).subscribe(res => {
+      if (useCache && me.caching) {
+        console.warn('Getting data from cache!')
+        let objects = me.queryFromCache(query)
+        resolve(objects)
+      }
+
+
+      me.query(query, useCache).pipe(take(1)).subscribe(res => {
 
         resolve(res.data)
 
@@ -359,9 +450,11 @@ export class BackendHttpServiceBase<T extends ObjectWithId> extends BackendServi
   }
 
 
-  async queryFirst$(query: DbQuery): Promise<T> {
+  async queryFirst$(query: DbQuery, useCache: boolean = true): Promise<T> {
 
-    let res = await this.query$(query)
+    // ObjectHelper.clone()
+    query.take = 1
+    let res = await this.query$(query, useCache)
 
     if (res?.length > 0)
       return res[0]
@@ -386,10 +479,15 @@ export class BackendHttpServiceBase<T extends ObjectWithId> extends BackendServi
    * 
    */
 
-  cache: T[]
-  cacheQuery: DbQuery
+  //cache: T[]
 
-  async initCache() {
+  cache: ObjectCache<T> = new ObjectCache<T>()
+  cacheQuery: DbQuery
+  caching: boolean = false
+  linkedTypes: string[] = []
+
+
+  async initCache(typeInfos: TypeInfo[]) {
 
     //this.cacheQuery = cacheQuery
 
@@ -397,10 +495,91 @@ export class BackendHttpServiceBase<T extends ObjectWithId> extends BackendServi
 
 
     if (!this.loadCacheFromDisk())
-      this.refreshCacheFromServer(this.cacheQuery)
+      await this.loadFullCacheFromBackend()
+    else
+      await this.updateCacheFromBackend(typeInfos)
+
+    this.caching = true
+
+  }
+
+  queryFromCache(query: DbQuery): T[] {
+
+    let count = 0
+
+    if (!this.cache)
+      throw new Error('Cache not initialised!')
+
+    let matches = this.cache.objects.filter(obj => {
+
+      if (count >= query.take)
+        return false
+
+      for (let condition of query.where.and) {
+        if (this.checkCondition(obj, condition) == false)
+          return false
+      }
+
+      count++
+      return true
+    })
+
+    if (ArrayHelper.AtLeastOneItem(query.order)) {
+
+      const orderFields = query.order.map(o => o.field)
+      const orders = query.order.map(o => o.order)
+      matches = _.orderBy(matches, orderFields, orders)
+
+    }
+
+    return matches
+
+  }
+
+  checkCondition(obj: ObjectWithId, condition: QueryCondition) {
+
+    const left = obj[condition.field]
+    let leftString
+
+    switch (condition.operator) {
+
+      case QueryOperator.equals:
+        return left == condition.value
+
+      case QueryOperator.not:
+        return left != condition.value
+
+      case QueryOperator.contains:
+        leftString = left as string
+        var regex = new RegExp(condition.value, "i");
+        return regex.test(leftString)
+
+      //return leftString.includes(condition.value)
+
+      case QueryOperator.startsWith:
+        leftString = left as string
+        return leftString.startsWith(condition.value)
+
+      case QueryOperator.endsWith:
+        leftString = left as string
+        return leftString.endsWith(condition.value)
+
+      case QueryOperator.in:
+
+        if (!ArrayHelper.AtLeastOneItem(condition.value))
+          return false
+
+        return (_.findIndex(condition.value, item => item == left) >= 0)
+      // condition.value.indexOf(left)
+
+
+      default:
+        throw new Error(`QueryOperator ${condition.operator} not implemented by cache`)
+    }
 
 
   }
+
 
   loadCacheFromDisk(): boolean {
 
@@ -414,10 +593,15 @@ export class BackendHttpServiceBase<T extends ObjectWithId> extends BackendServi
       //   const byteArray = this.str2ab(cacheString)
       //   const json = '' // this.decompress(byteArray)
       const json = cacheString
-      const objects = JSON.parse(json)
-      const typedData = objects.map(obj => plainToInstance(this.type, obj))
+      const cache = JSON.parse(json)
 
-      this.cache = typedData
+      const typedCache = plainToInstance(ObjectCache<T>, cache)
+      typedCache.objects = typedCache.objects.map(obj => plainToInstance(this.type, obj))
+      // const typedData = objects.map(obj => plainToInstance(this.type, obj))
+
+      console.warn('Cache loaded from disk', typedCache)
+
+      this.cache = typedCache
 
       return true
     } catch (error) {
@@ -426,22 +610,153 @@ export class BackendHttpServiceBase<T extends ObjectWithId> extends BackendServi
       return false
 
     }
+  }
 
 
+  async updateCache(objectChange: ObjectChange<T>) {
+
+    if (!this.caching)
+      return
+
+    console.error('Update cache for change!!', objectChange)
+
+    //if (objectChange.change == ObjectChangeType.)
+
+    let objId = objectChange.objectId()
+
+    switch (objectChange.change) {
+      case ObjectChangeType.create:
+        await this.refreshCachedObjectFromBackend(objId)
+        this.writeCacheToStorage()
+
+        break
+
+      case ObjectChangeType.update:
+      case ObjectChangeType.softDelete:
+        {
+          if (!objId)
+            return
+
+          const cachedObject = this.cache.objects.find(obj => obj.id == objId)
+          const updates = objectChange.object
+          delete updates['id']
+
+          if (!cachedObject)
+            throw new Error(`Can't update cache, because object not in cache (id=${objId})`)
+
+          Object.keys(updates).forEach(property => {
+
+            cachedObject[property] = updates[property]
+          })
+          this.writeCacheToStorage()
+
+        }
+        break
+
+    }
+  }
+
+
+  /** refresh an individual object in the cache & saves cache to storage */
+  async refreshCachedObjectFromBackend(id: string): Promise<T> {
+
+    let qry = new DbQuery()
+    qry.and('id', QueryOperator.equals, id)
+    qry.includes = this.cacheQuery.includes  // include all necessary linked objects
+
+    const obj = await this.queryFirst$(qry, false)
+
+    console.error(obj)
+
+    let idx = _.findIndex(this.cache.objects, obj => obj.id == id)
+
+    if (obj) {
+
+      if (idx >= 0) // remove & replace item in cache
+        this.cache.objects.splice(idx, 1, obj)
+      else
+        this.cache.objects.push(obj)
+
+    } else {
+
+      if (idx >= 0) // object not found => remove item from cache
+        this.cache.objects.splice(idx, 1)
+    }
+
+    this.writeCacheToStorage()
+
+    return obj
 
   }
 
 
-  async refreshCacheFromServer(dbQuery: DbQuery) {
+  async updateCacheFromBackend(typeInfos: TypeInfo[]) {
 
-    this.cache = await this.query$(dbQuery)
 
-    let json = JSON.stringify(this.cache)
+    let now = new Date()
+
+    this.cache.lastSync
+
+    const typeInfo = typeInfos.find(ti => ti.name == this.typeName)
+
+    if (!typeInfo) {
+      console.log(`Can not update cache from backend: no typeInfo existing for ${this.typeName}`)
+      return
+    }
+
+    let lastUpdated: Date
+
+    if (typeInfo) {
+      lastUpdated = typeInfo.lastUpdated
+    }
+
+    if (lastUpdated && lastUpdated > this.cache.lastSync) {
+      let updatedObjects = await this.getUpdatedObjects(this.cache.lastSync)
+
+      console.warn('Updated CACHE objects', updatedObjects)
+
+      if (this.cache.updateObjectsInCache(updatedObjects)) {
+        this.cache.lastSync = now
+        this.writeCacheToStorage()
+      }
+
+    }
+  }
+
+
+
+  async getUpdatedObjects(since: Date): Promise<T[]> {
+    const qry = new DbQuery()
+    qry.or('upd', QueryOperator.greaterThanOrEqual, since)
+    qry.includes = this.cacheQuery.includes
+
+    const objects = await this.query$(qry, false)
+
+    return objects
+  }
+
+  async loadFullCacheFromBackend() {
+
+    let me = this
+
+    me.cache.objects = await me.query$(this.cacheQuery, false)
+    me.cache.lastSync = new Date()
+
+    //console.warn(me.cache)
+
+    this.writeCacheToStorage()
+
+  }
+
+  writeCacheToStorage() {
+    let me = this
+
+    let json = JSON.stringify(me.cache)
 
     let cacheString = json  // await this.compressToString(json)
-    localStorage.setItem(this.urlDifferentiator, cacheString)
-
+    localStorage.setItem(me.urlDifferentiator, cacheString)
   }
+
 
   async compressToString(
     str: string,
